Made significant changes to the annotateAndAddVariant function.
converted all of the execAsync to either spawn() or fork() to more appropriately spawn a new thread instead of clogging the current thread. Additionally, am no longer calling parseVCF directly but forking to it and passing json data to it. This alleviates stress on the server.

Secondly Added a test to ensure the number of documents entered into the database coincides with the number of entries within the actual file.