Made major changes to the parseVCF function.
It no longer is an exportable module, but is callable via the command line, taking in 2 arguments, the file name and a json string of patient data. It now opens a readStream that it monitors for incoming data event. when data comes in it parses it and when a threshold of documents have been reached, it inserts them into the mongodb.

It prints an array with values that were ignored to the console for retrieval